# src/main/resources/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: API Workshop
  version: "1.0"
  description: Documentación del laboratorio

tags:
  - name: workshop
    description: Endpoints de salud y saludos
  - name: simulaciones
    description: Simulador de préstamo
  - name: demo-estado
    description: Demostración singleton vs instancia manual

paths:

  /api/v1:
    get:
      tags: [workshop]
      operationId: getWorkshopHealth
      summary: Health check del workshop
      responses:
        "200":
          description: Servicio activo
          content:
            application/json:
              schema:
                type: object
                properties:
                  estado:
                    type: string
                    example: ok
                  mensaje:
                    type: string
                    example: Workshop Spring Boot activo

  /api/v1/saludos:
    get:
      tags: [workshop]
      operationId: saludarPorGet
      summary: Saludo por GET con nombre opcional
      parameters:
        - name: nombre
          in: query
          required: false
          schema:
            type: string
            default: Mundo
      responses:
        "200":
          description: Saludo generado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SaludoResponse'

    post:
      tags: [workshop]
      operationId: saludarPorPost
      summary: Saludo por POST con cuerpo JSON
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaludoRequest'
      responses:
        "200":
          description: Saludo generado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SaludoResponse'
        "400":
          description: Validación fallida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/simulaciones/prestamo:
    post:
      tags: [simulaciones]
      operationId: simularPrestamo
      summary: Simula las cuotas de un préstamo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PrestamoRequest'
      responses:
        "200":
          description: Simulación exitosa
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrestamoResponse'
        "400":
          description: Validación fallida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:

    SaludoRequest:
      type: object
      required: [nombre]
      properties:
        nombre:
          type: string
          minLength: 1
          example: Ana

    SaludoResponse:
      type: object
      properties:
        mensaje:
          type: string
          example: Hola Ana

    PrestamoRequest:
      type: object
      required: [monto, tasaAnual, meses]
      properties:
        monto:
          type: number
          minimum: 0.01
          example: 10000
        tasaAnual:
          type: number
          minimum: 0.01
          example: 12
        meses:
          type: integer
          minimum: 1
          maximum: 360
          example: 12

    PrestamoResponse:
      type: object
      properties:
        cuotaMensual:
          type: number
          example: 888.49
        interesTotal:
          type: number
          example: 661.85
        totalPagar:
          type: number
          example: 10661.85

    ErrorResponse:
      type: object
      properties:
        codigo:
          type: string
          example: VALIDATION_ERROR
        mensaje:
          type: string
          example: El nombre es obligatorio

    EstadoResponse:
      type: object
      properties:
        tipo:
          type: string
          example: singleton
        valorActual:
          type: integer
          example: 25

  /api/v1/demo/estado/singleton/{valor}:
    post:
      tags: [demo-estado]
      operationId: actualizarSingleton
      summary: Actualiza el valor del bean singleton
      parameters:
        - name: valor
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Valor actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EstadoResponse'

  /api/v1/demo/estado/singleton:
    get:
      tags: [demo-estado]
      operationId: obtenerSingleton
      summary: Obtiene el valor actual del bean singleton
      responses:
        "200":
          description: Valor actual
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EstadoResponse'

  /api/v1/demo/estado/singleton/reset:
    post:
      tags: [demo-estado]
      operationId: reiniciarSingleton
      summary: Reinicia el valor del bean singleton a 0
      responses:
        "200":
          description: Valor reiniciado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EstadoResponse'

  /api/v1/demo/estado/manual/{valor}:
    post:
      tags: [demo-estado]
      operationId: actualizarManual
      summary: Actualiza el valor instanciando el servicio manualmente (new)
      parameters:
        - name: valor
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Valor actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EstadoResponse'

  /api/v1/demo/estado/manual:
    get:
      tags: [demo-estado]
      operationId: obtenerManual
      summary: Obtiene el valor usando instancia manual (siempre 0)
      responses:
        "200":
          description: Valor actual
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EstadoResponse'